Make random string generation efficient
The current algorithm requires calling the libsodium random function 'size' times where size is the expected length of random string
Changed this so that this will require at the most '(size/64)' number of calls.
Also we do not need the chars list now we always use only alphabets and digits available in hex this makes it very easy and still gives us what we need.
Every instance of usage of this function has been checked and nowhere the 'chars' parameter is used