INDY-1112: change primeries election procedure for backup instances. (#539)
* INDY-1112: change primeries election procedure for backup instances.

Now primaries for backup instances are choosen in round-robin
manner always starting from primary. If the next node is a primary
for some instance then this node is skipped. So the first non-primary
node is choosen as primary for current instance. Such approach allows
to avoid election of instances of the same node as a primeries for
different instances.
The election procedure of the primary for master instance is not changed.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>

* Add some stylistical fixes, add comments.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>

* Add test for primary demotion and promotion.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>

* Add test for primaries selection routines.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>

* Re-factor primary selector tests.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>

* Remove adding of node itself to nodeReg during initialisation of txnPoolManager.

Now a node adds itself to nodeReg during catch-up of pool ledger.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>

* Fix test.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>

* Update node's HA in node registry on pool ledger catch-up reply.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>

* Add separate checks for IP/port to be updated for HA and cliHA.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>