Fix apply() function of domain request handler.
The appendTxns() function of ledger class have a side effect:
it appends seqNo to the passed txn object. But this object may
be a copy of original txn object. So if further processing of
original txn object requires seqNo then it will fail.
So now seqNo is appended to initial transaction object if the
copy was created.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>