INDY-1390: reduce MAX_RECONNECT_RETRY_ON_SAME_SOCKET from 5 to 1.
Trully speaking actually we don't fully understand the issues with
connections of newly added nodes. But tests and logs show that we
have troubles with connections to/from newly added nodes when we
try to use the same socket for sending of pings before other nodes
know about the node to be added to the pool.

So more investigation is needed but this changes may be applied as
a hot fix.

Signed-off-by: Sergey Shilov <sergey.shilov@dsr-company.com>